requirements.txt
ibapi==9.81.1.post1
flask==2.3.3
flask-cors==4.0.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly==5.15.0
python-dotenv==1.0.0
websocket-client==1.6.1
requests==2.31.0
asyncio-mqtt==0.13.0
schedule==1.2.0
